Fix: MainPage Video Error : choongnyeong
